Revert "xiugai report api return"
This reverts commit bb6c5c90473a8bb25faae445cad6452c157f4c4e.

+++ cn.jpush.api/report/ReportClient.cs
@@ -71,7 +71,7 @@ namespace cn.jpush.api.report
                     string trim = s.Trim();
                     if (!string.IsNullOrEmpty(trim))
                     {
-                        long.Parse(trim);
+                        int.Parse(trim);
                         list.Add(trim);
                     }
                 }
